+++ src/Cand3s.c
@@ -183,6 +183,9 @@ SEXP Cands(SEXP oo1, SEXP xx1, SEXP yy1,
 
   const int o1 = sex2op(oo1);
   const int o2 = sex2op(oo2);
+  if (o1 == 0 || (use2 && o2 == 0)) {
+    return R_NilValue;
+  }
   SEXP ans = PROTECT(allocVector(RAWSXP, N));
   unsigned char * ansp = RAW(ans);
   int err[1] = {0};
@@ -235,7 +238,6 @@ SEXP Cands(SEXP oo1, SEXP xx1, SEXP yy1,
   }
   UNPROTECT(1);
   if (err[0]) {
-    REprintf("Unsupported type\n");
     return R_NilValue;
   }
   return ans;